Rewrite Project on TypeScript
