+++ src/components/redux/store.ts
@@ -7,4 +7,6 @@ const store = configureStore({
 	},
 });
 
+export type RootState = ReturnType<typeof store.getState>;
+
 export default store;
